♻️ refactor : interface 파일 통합
