+++ src/api/index.ts
@@ -15,7 +15,6 @@ apiClient.interceptors.response.use(
     if (error.response.status === 403) {
       try {
         const res = await postRefreshToken();
-        // console.log("응닫받은거", res);
         const newAccessToken = res.data.accessToken;
         const grantType = res.data.grantType;
