+++ src/components/MyPage/EmptyMessage.tsx
@@ -1,8 +1,4 @@
-interface EmptyMessageProps {
-  message: string;
-}
-
-const EmptyMessage: React.FC<EmptyMessageProps> = ({ message }) => {
+const EmptyMessage = ({ message }: EmptyMessageProps) => {
   return (
     <div className="flex justify-center mt-[20%]">
       <h3 className="text-smTitle">{message}</h3>
